# Conflicts:
#	.gitignore
#	src/Cool/Log/ToUser.cpp
#	src/Cool/Log/internal/Message.h
#	src/Cool/MessageConsole/MessageConsole.cpp
#	src/Cool/MessageConsole/MessageWithMetadata.h

+++ src/Cool/MessageConsole/MessageWithMetadata.h
@@ -7,7 +7,7 @@ namespace Cool::internal {
 struct MessageWithMetadata {
     Message                               message;
     std::chrono::system_clock::time_point timestamp;
-    uint64_t                              count;
+    uint32_t                              count;
     bool                                  forced_to_be_closable = false;
 };
